74 - tidy up
